Rename NewParameter[Frame] -> Parameter[Frame]
Now the old versions of the classes have been removed, we can rename the
new ones.
