+++ bluemira/base/solver.py
@@ -49,7 +49,7 @@ import abc
 import enum
 from typing import Any, Callable, Optional, Type
 
-from bluemira.base.parameter_frame import NewParameterFrame as ParameterFrame
+from bluemira.base.parameter_frame import ParameterFrame
 
 
 class Task(abc.ABC):
