+++ bluemira/codes/process/_solver.py
@@ -27,8 +27,7 @@ from typing import Dict, List, Tuple, Union
 import numpy as np
 
 from bluemira.base.look_and_feel import bluemira_warn
-from bluemira.base.parameter_frame import NewParameterFrame as ParameterFrame
-from bluemira.base.parameter_frame import make_parameter_frame
+from bluemira.base.parameter_frame import ParameterFrame, make_parameter_frame
 from bluemira.base.solver import RunMode as BaseRunMode
 from bluemira.codes.error import CodesError
 from bluemira.codes.interface import CodesSolver
