+++ bluemira/magnetostatics/greens.py
@@ -101,7 +101,10 @@ def circular_coil_inductance_elliptic(radius, rc):
         The self-inductance of the circular coil [H]
     """
     k = 4 * radius * (radius - rc) / (2 * radius - rc) ** 2
-    k = clip_nb(k, GREENS_ZERO, 1.0 - GREENS_ZERO)
+    args = (
+        np.array(arg, dtype=np.float64) for arg in (k, GREENS_ZERO, 1.0 - GREENS_ZERO)
+    )
+    k = clip_nb(*args)
     return MU_0 * (2 * radius - rc) * ((1 - k**2 / 2) * ellipk(k) - ellipe(k))
