+++ tests/codes/test_interface.py
@@ -21,7 +21,7 @@
 
 from dataclasses import dataclass
 
-from bluemira.base.parameter_frame import NewParameter as Parameter
+from bluemira.base.parameter_frame import Parameter as Parameter
 from bluemira.base.solver import NoOpTask, RunMode
 from bluemira.codes import interface
 from bluemira.codes.params import MappedParameterFrame, ParameterMapping
